requirements.txt
json2html==1.3.0
ping3==4.0.8
PySide6==6.7.1
PySide6_Addons==6.7.1
PySide6_Essentials==6.7.1
python_nmap==0.7.1
shodan==1.31.0